% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gating_plot.R
\name{build_gating_plot}
\alias{build_gating_plot}
\title{build_gating_plot}
\usage{
build_gating_plot(
  sample,
  n_vars,
  var_x,
  var_y,
  gating_type,
  slider_1v = NULL,
  side_qnt_1v = "left",
  slider_x = NULL,
  slider_y = NULL,
  ellipse_val = NULL,
  apply_gate = FALSE
)
}
\arguments{
\item{sample}{A flowFrame object.}

\item{n_vars}{Character, either 'one variable' or 'two variables'.}

\item{var_x}{Character, the name of the X-axis channel.}

\item{var_y}{Character, the name of the Y-axis channel (optional for 1D).}

\item{gating_type}{Character, 'Rectangular' or 'Ellipse' (for 2D).}

\item{slider_1v}{Numeric, quantile value for 1D gating.}

\item{side_qnt_1v}{Character, either 'left' or 'right' corresponding to the side of cut.}

\item{slider_x}{Numeric vector of length 2 for rectangular X limits.}

\item{slider_y}{Numeric vector of length 2 for rectangular Y limits.}

\item{ellipse_val}{Numeric, quantile for flowClust 2D gating.}

\item{apply_gate}{Logical, if TRUE apply the gate and compute the plot.}
}
\value{
A ggplot or ggcyto object.
}
\description{
Generate Gating Plots for Flow Cytometry
}
