% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_workflow_log.R
\name{save_workflow_log}
\alias{save_workflow_log}
\title{save_workflow_log}
\usage{
save_workflow_log(output_path, peac_df, gat_df, filename = "workflow_log.txt")
}
\arguments{
\item{output_path}{Character. The directory where the log file will be saved.}

\item{peac_df}{Data frame. The PeacoQC results to include in the log.}

\item{gat_df}{Data frame. The Gating results to include in the log.}

\item{filename}{Character. The name of the log file. Default is "workflow_log.txt".}
}
\description{
This function saves a single text file containing both PeacoQC and Gating information. Each section is clearly labeled with a title for readability.
}
