% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendmessages.R
\name{sendmessages}
\alias{sendmessages}
\title{sendmessages}
\usage{
sendmessages(text, type, duration = 4)
}
\arguments{
\item{text}{A string to be printed.}

\item{type}{Type of the message between "info", "warning", "danger", "success", and "gears".}

\item{duration}{Number of seconds to display the message before it disappears. Use NULL to make the message not automatically disappear.}
}
\description{
\code{sendmessages} Send a message in console and a notification in shiny if it's running.
}
